--- docs/rules/a11y-aria-label-is-readable.md
@@ -1,18 +1,0 @@
-# [aria-label] text should be formatted as you would visible text, in a human-readable format (`github/a11y-aria-label-is-readable`)
-
-💼 This rule is enabled in the ⚛️ `react` config.
-
-<!-- end auto-generated rule header -->
-
-## Rule Details
-
-## Resources
-
-## Examples
-
-### **Incorrect** code for this rule 👎
-
-### **Correct** code for this rule 👍
-
-
-## Version
