+++ lib/index.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: {
     'a11y-no-generic-link-text': require('./rules/a11y-no-generic-link-text'),
-    'a11y-aria-label-is-readable': require('./rules/a11y-aria-label-is-readable'),
+    'a11y-aria-label-is-well-formatted': require('./rules/a11y-aria-label-is-well-formatted'),
     'array-foreach': require('./rules/array-foreach'),
     'async-currenttarget': require('./rules/async-currenttarget'),
     'async-preventdefault': require('./rules/async-preventdefault'),
